Sharing common code between SAML and OIDC Auth Configuration - moving common code to ManageIQ::Appliance::AuthUtilities
