+++ lib/manageiq/appliance_console/saml_authentication.rb
@@ -1,11 +1,10 @@
-require "uri"
-
 module ManageIQ
   module ApplianceConsole
     class SamlAuthentication
+      include ManageIQ::ApplianceConsole::AuthUtilities
+
       MELLON_CREATE_METADATA_COMMAND = Pathname.new("/usr/libexec/mod_auth_mellon/mellon_create_metadata.sh")
 
-      HTTPD_CONFIG_DIRECTORY = Pathname.new("/etc/httpd/conf.d")
       SAML2_CONFIG_DIRECTORY = Pathname.new("/etc/httpd/saml2")
       IDP_METADATA_FILE      = SAML2_CONFIG_DIRECTORY.join("idp-metadata.xml")
 
@@ -105,14 +104,6 @@ module ManageIQ
         HTTPD_CONFIG_DIRECTORY.join("manageiq-external-auth-saml.conf").exist?
       end
 
-      def restart_httpd
-        httpd_service = LinuxAdmin::Service.new("httpd")
-        if httpd_service.running?
-          say("Restarting httpd ...")
-          httpd_service.restart
-        end
-      end
-
       # SAML IDP Metadata
 
       def validate_saml_idp_metadata_option
@@ -122,29 +113,7 @@ module ManageIQ
         raise "Missing SAML IDP metadata file #{idp_metadata}" if path_is_file?(idp_metadata) && !File.exist?(idp_metadata)
       end
 
-      def path_is_file?(path)
-        path.present? && !path_is_url?(path)
-      end
-
-      def path_is_url?(path)
-        path =~ /\A#{URI.regexp(["http", "https"])}\z/x
-      end
-
       # File Management
-
-      def remove_file(path)
-        if path.exist?
-          debug_msg("Removing #{path} ...")
-          File.delete(path)
-        end
-      end
-
-      def copy_template(dir, file)
-        src_path = template_directory.join(relative_from_root(dir), file)
-        dest_path = dir.join(file)
-        debug_msg("Copying template #{src_path} to #{dest_path} ...")
-        FileUtils.cp(src_path, dest_path)
-      end
 
       def download_network_file(source_file_url, target_file)
         require "net/http"
@@ -154,14 +123,6 @@ module ManageIQ
         raise "Failed to download file from #{source_file_url}" unless result.kind_of?(Net::HTTPSuccess)
 
         File.write(target_file, result.body)
-      end
-
-      def template_directory
-        @template_directory ||= Pathname.new(ENV.fetch("APPLIANCE_TEMPLATE_DIRECTORY"))
-      end
-
-      def relative_from_root(path)
-        path.absolute? ? path.relative_path_from(Pathname.new("/")) : path
       end
 
       # Appliance Settings
@@ -177,31 +138,6 @@ module ManageIQ
           "/authentication/provider_type=saml"
         ]
         Utilities.rake_run("evm:settings:set", params)
-      end
-
-      def configure_auth_settings_database
-        say("Setting Appliance Authentication Settings to Database ...")
-        params = [
-          "/authentication/mode=database",
-          "/authentication/httpd_role=false",
-          "/authentication/saml_enabled=false",
-          "/authentication/oidc_enabled=false",
-          "/authentication/sso_enabled=false",
-          "/authentication/provider_type=none"
-        ]
-        Utilities.rake_run("evm:settings:set", params)
-      end
-
-      # Logging
-
-      def debug_msg(msg)
-        say(msg) if options[:verbose]
-      end
-
-      def log_command_error(err)
-        say(err.result.output)
-        say(err.result.error)
-        say("")
       end
     end
   end
